requirements.txt
openai-whisper
torch
transformers
camel-tools
google-cloud-texttospeech
sounddevice
soundfile
gtts
